Fix integration with storysource
