+++ .storybook/webpack.config.js
@@ -13,6 +13,12 @@ module.exports = async ({ config, mode }) => {
     include: path.resolve(__dirname, '../')
   });
 
+  config.module.rules.push({
+    test: /\.stories\.jsx?$/,
+    loaders: [require.resolve('@storybook/addon-storysource/loader')],
+    enforce: 'pre'
+  });
+
   // Return the altered config
   return config;
 };
