+++ package.json
@@ -44,6 +44,7 @@
     "@babel/preset-react": "^7.0.0",
     "@storybook/addon-actions": "^5.0.10",
     "@storybook/addon-links": "^5.0.10",
+    "@storybook/addon-storysource": "^5.0.10",
     "@storybook/addons": "^5.0.10",
     "@storybook/react": "^5.0.10",
     "@storybook/theming": "^5.0.10",
