+++ src/__stories__/01-basic-table/00-Basic.stories.js
@@ -1,10 +1,12 @@
 import React from 'react'; // Import React
+import { storiesOf } from '@storybook/react';
+import { categoryName } from './_base';
 
 import moment from 'moment'; // Example for onSort prop
 import Datatable from '../../Table'; // Import this package
 
 const header = [
-  { title: 'Username (filterable)', prop: 'username', filterable: true },
+  { title: 'Username', prop: 'username' },
   { title: 'Name', prop: 'realname' },
   { title: 'Location', prop: 'location' },
   { title: 'Last Updated', prop: 'date' }
@@ -34,15 +36,10 @@ const body = Array.from(new Array(57), () => {
   };
 });
 
-const Filter = {
-  name: 'Filter',
-  story: () => (
-    <Datatable
-      tableHeader={header}
-      tableBody={body}
-      tableClass="striped hover responsive"
-    />
-  )
-};
-
-export default Filter;
+storiesOf(categoryName, module).add('Basic', () => (
+  <Datatable
+    tableHeader={header}
+    tableBody={body}
+    tableClass="striped hover responsive"
+  />
+));
