+++ src/__stories__/01-basic-table/02-Sort.stories.js
@@ -1,21 +1,13 @@
 import React from 'react'; // Import React
+import { storiesOf } from '@storybook/react';
+import { categoryName } from './_base';
 
 import moment from 'moment'; // Example for onSort prop
 import Datatable from '../../Table'; // Import this package
 
 const header = [
-  {
-    title: 'Username (filterable)',
-    prop: 'username',
-    sortable: true,
-    filterable: true
-  },
+  { title: 'Username', prop: 'username', sortable: true },
   { title: 'Name', prop: 'realname', sortable: true },
-  {
-    title: 'Name Uppercased',
-    prop: 'realnameuppercase',
-    cell: row => row.realname.toUpperCase()
-  },
   { title: 'Location', prop: 'location' },
   { title: 'Last Updated', prop: 'date', sortable: true }
 ];
@@ -52,30 +44,12 @@ const onSortFunction = {
   }
 };
 
-const customLabels = {
-  first: '<<',
-  last: '>>',
-  prev: '<',
-  next: '>',
-  show: 'Display',
-  entries: 'rows',
-  noResults: 'There is no data to be displayed'
-};
-
-const CustomCell = {
-  name: 'Custom Cell Representation',
-  story: () => (
-    <Datatable
-      tableHeader={header}
-      tableBody={body}
-      tableClass="striped hover responsive"
-      rowsPerPage={5}
-      rowsPerPageOption={[5, 10, 15, 20]}
-      initialSort={{ prop: 'username', isAscending: true }}
-      onSort={onSortFunction}
-      labels={customLabels}
-    />
-  )
-};
-
-export default CustomCell;
+storiesOf(categoryName, module).add('Sort', () => (
+  <Datatable
+    tableHeader={header}
+    tableBody={body}
+    tableClass="striped hover responsive"
+    initialSort={{ prop: 'username', isAscending: true }}
+    onSort={onSortFunction}
+  />
+));
