+++ src/__stories__/01-basic-table/05-CustomLabels.stories.js
@@ -1,4 +1,6 @@
 import React from 'react'; // Import React
+import { storiesOf } from '@storybook/react';
+import { categoryName } from './_base';
 
 import moment from 'moment'; // Example for onSort prop
 import Datatable from '../../Table'; // Import this package
@@ -47,19 +49,25 @@ const onSortFunction = {
   }
 };
 
-const Combination = {
-  name: 'Filter, sort, pagination',
-  story: () => (
-    <Datatable
-      tableHeader={header}
-      tableBody={body}
-      tableClass="striped hover responsive"
-      rowsPerPage={5}
-      rowsPerPageOption={[5, 10, 15, 20]}
-      initialSort={{ prop: 'username', isAscending: true }}
-      onSort={onSortFunction}
-    />
-  )
+const customLabels = {
+  first: '<<',
+  last: '>>',
+  prev: '<',
+  next: '>',
+  show: 'Display',
+  entries: 'rows',
+  noResults: 'There is no data to be displayed'
 };
 
-export default Combination;
+storiesOf(categoryName, module).add('Custom Labels', () => (
+  <Datatable
+    tableHeader={header}
+    tableBody={body}
+    tableClass="striped hover responsive"
+    rowsPerPage={5}
+    rowsPerPageOption={[5, 10, 15, 20]}
+    initialSort={{ prop: 'username', isAscending: true }}
+    onSort={onSortFunction}
+    labels={customLabels}
+  />
+));
