+++ src/__stories__/01-basic-table/_base.js
@@ -1,0 +1,1 @@
+export const categoryName = 'Basic Table';
